requirements.txt
# Core web framework and utilities
Flask==2.3.3
Flask-SocketIO==5.3.6
eventlet==0.33.3
python-socketio==5.10.0
python-engineio==4.8.0
requests==2.31.0

# Data handling and analysis
pandas==2.1.1
numpy==1.24.3
yfinance==0.2.31

# Technical Analysis (Production only)
# ta-lib==0.4.28  # Uncomment for production deployment

# API and data sources
alpha_vantage==2.3.1
python-dotenv==1.0.0

# Database
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.20

# Authentication and security
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0

# PDF generation
reportlab==4.0.4

# Date and time handling
pytz==2023.3
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0
Werkzeug==2.3.7

# Caching
Flask-Caching==2.1.0